Move the aggregation code to istore_type.c.template

+++ src/avl.c
@@ -35,7 +35,7 @@ istore_make_empty(AvlNode *t)
  * find a node for a given key
  */
 AvlNode *
-is_tree_find(int32 key, AvlNode *t)
+is_tree_find(int64 key, AvlNode *t)
 {
     int32 cmp;
 
@@ -142,7 +142,7 @@ is_tree_length(AvlNode *t)
  * insert key/value into tree
  */
 AvlNode *
-is_tree_insert(AvlNode *t, int32 key, int64 value)
+is_tree_insert(AvlNode *t, int64 key, int64 value)
 {
     if (t == NULL)
     {
