+++ src/avl.h
@@ -2,14 +2,13 @@
 #define ISTORE_AVL_H
 
 #include "postgres.h"
-#include "utils/int8.h"
 
 
 typedef struct AvlNode AvlNode;
 
 struct AvlNode
 {
-    int32    key;
+    int64    key;
     int64    value;
     AvlNode  *left;
     AvlNode  *right;
@@ -17,7 +16,7 @@ struct AvlNode
 };
 
 AvlNode* istore_make_empty(AvlNode *t);
-AvlNode* is_tree_find(int32 key, AvlNode *t);
-AvlNode* is_tree_insert(AvlNode *t, int32 key, int64 value);
+AvlNode* is_tree_find(int64 key, AvlNode *t);
+AvlNode* is_tree_insert(AvlNode *t, int64 key, int64 value);
 int is_tree_length(AvlNode* t);
 #endif // ISTORE_AVL_H
